+++ spinta/config.py
@@ -44,6 +44,8 @@ CONFIG = {
             'xlsx': 'spinta.manifests.tabular.components:XlsxManifest',
             'rdf': 'spinta.manifests.rdf.components:RdfManifest',
             'gsheets': 'spinta.manifests.tabular.components:GsheetsManifest',
+            'sql': 'spinta.manifests.sql.components:SqlManifest',
+            'memory': 'spinta.manifests.memory.components:MemoryManifest'
         },
         'backends': {
             # In memory backends mostly usable in tests